+++ aiobotocore/credentials.py
@@ -1,11 +1,8 @@
 import asyncio
-import datetime
-import json
 import logging
 import os
 import subprocess
 from copy import deepcopy
-from hashlib import sha1
 
 import botocore.compat
 from botocore import UNSIGNED
@@ -27,6 +24,7 @@ from botocore.credentials import (
     CredentialResolver,
     CredentialRetrievalError,
     Credentials,
+    DeferredRefreshableCredentials,
     EnvProvider,
     InstanceMetadataProvider,
     InvalidConfigError,
@@ -39,6 +37,7 @@ from botocore.credentials import (
     RefreshableCredentials,
     RefreshWithMFAUnsupportedError,
     SharedCredentialProvider,
+    SSOCredentialFetcher,
     SSOProvider,
     SSOTokenLoader,
     UnauthorizedSSOTokenError,
@@ -50,7 +49,6 @@ from botocore.credentials import (
     parse,
     resolve_imds_endpoint_mode,
 )
-from dateutil.tz import tzutc
 
 from aiobotocore._helpers import resolve_awaitable
 from aiobotocore.config import AioConfig
@@ -251,7 +249,7 @@ create_aio_mfa_serial_refresher = create_mfa_serial_refresher
 class AioCredentials(Credentials):
     async def get_frozen_credentials(self):
         return ReadOnlyCredentials(
-            self.access_key, self.secret_key, self.token
+            self.access_key, self.secret_key, self.token, self.account_id
         )
 
 
@@ -301,6 +299,19 @@ class AioRefreshableCredentials(RefreshableCredentials):
     def token(self, value):
         self._token = value
 
+    @property
+    def account_id(self):
+        # TODO: this needs to be resolved
+        raise NotImplementedError(
+            "missing call to self._refresh. "
+            "Use get_frozen_credentials instead"
+        )
+        return self._account_id
+
+    @account_id.setter
+    def account_id(self, value):
+        self._account_id = value
+
     async def _refresh(self):
         if not self.refresh_needed(self._advisory_refresh_timeout):
             return
@@ -349,7 +360,7 @@ class AioRefreshableCredentials(RefreshableCredentials):
             return
         self._set_from_data(metadata)
         self._frozen_credentials = ReadOnlyCredentials(
-            self._access_key, self._secret_key, self._token
+            self._access_key, self._secret_key, self._token, self._account_id
         )
         if self._is_expired():
             msg = (
@@ -364,22 +375,20 @@ class AioRefreshableCredentials(RefreshableCredentials):
         return self._frozen_credentials
 
 
-class AioDeferredRefreshableCredentials(AioRefreshableCredentials):
+class AioDeferredRefreshableCredentials(
+    DeferredRefreshableCredentials, AioRefreshableCredentials
+):
     def __init__(self, refresh_using, method, time_fetcher=_local_now):
         self._refresh_using = refresh_using
         self._access_key = None
         self._secret_key = None
         self._token = None
+        self._account_id = None
         self._expiry_time = None
         self._time_fetcher = time_fetcher
         self._refresh_lock = asyncio.Lock()
         self.method = method
         self._frozen_credentials = None
-
-    def refresh_needed(self, refresh_in=None):
-        if self._frozen_credentials is None:
-            return True
-        return super().refresh_needed(refresh_in)
 
 
 class AioCachedCredentialFetcher(CachedCredentialFetcher):
@@ -404,12 +413,15 @@ class AioCachedCredentialFetcher(CachedCredentialFetcher):
 
         creds = response['Credentials']
         expiration = _serialize_if_needed(creds['Expiration'], iso=True)
-        return {
+        credentials = {
             'access_key': creds['AccessKeyId'],
             'secret_key': creds['SecretAccessKey'],
             'token': creds['SessionToken'],
             'expiry_time': expiration,
+            'account_id': creds.get('AccountId'),
         }
+
+        return credentials
 
 
 class AioBaseAssumeRoleCredentialFetcher(
@@ -426,7 +438,9 @@ class AioAssumeRoleCredentialFetcher(
         kwargs = self._assume_role_kwargs()
         client = await self._create_client()
         async with client as sts:
-            return await sts.assume_role(**kwargs)
+            response = await sts.assume_role(**kwargs)
+            self._add_account_id_to_response(response)
+            return response
 
     async def _create_client(self):
         """Create an STS client using the source credentials."""
@@ -470,7 +484,9 @@ class AioAssumeRoleWithWebIdentityCredentialFetcher(
         # the token, explicitly configure the client to not sign requests.
         config = AioConfig(signature_version=UNSIGNED)
         async with self._client_creator('sts', config=config) as client:
-            return await client.assume_role_with_web_identity(**kwargs)
+            response = await client.assume_role_with_web_identity(**kwargs)
+            self._add_account_id_to_response(response)
+            return response
 
     def _assume_role_kwargs(self):
         """Get the arguments for assume role based on current configuration."""
@@ -503,6 +519,7 @@ class AioProcessProvider(ProcessProvider):
             secret_key=creds_dict['secret_key'],
             token=creds_dict.get('token'),
             method=self.METHOD,
+            account_id=creds_dict.get('account_id'),
         )
 
     async def _retrieve_credentials_using(self, credential_process):
@@ -533,6 +550,7 @@ class AioProcessProvider(ProcessProvider):
                 'secret_key': parsed['SecretAccessKey'],
                 'token': parsed.get('SessionToken'),
                 'expiry_time': parsed.get('Expiration'),
+                'account_id': self._get_account_id(parsed),
             }
         except KeyError as e:
             raise CredentialRetrievalError(
@@ -578,6 +596,7 @@ class AioEnvProvider(EnvProvider):
                     expiry_time,
                     refresh_using=fetcher,
                     method=self.METHOD,
+                    account_id=credentials['account_id'],
                 )
 
             return AioCredentials(
@@ -585,6 +604,7 @@ class AioEnvProvider(EnvProvider):
                 credentials['secret_key'],
                 credentials['token'],
                 method=self.METHOD,
+                account_id=credentials['account_id'],
             )
         else:
             return None
@@ -626,8 +646,13 @@ class AioSharedCredentialProvider(SharedCredentialProvider):
                     config, self.ACCESS_KEY, self.SECRET_KEY
                 )
                 token = self._get_session_token(config)
+                account_id = self._get_account_id(config)
                 return AioCredentials(
-                    access_key, secret_key, token, method=self.METHOD
+                    access_key,
+                    secret_key,
+                    token,
+                    method=self.METHOD,
+                    account_id=account_id,
                 )
 
 
@@ -648,8 +673,13 @@ class AioConfigProvider(ConfigProvider):
                     profile_config, self.ACCESS_KEY, self.SECRET_KEY
                 )
                 token = self._get_session_token(profile_config)
+                account_id = self._get_account_id(profile_config)
                 return AioCredentials(
-                    access_key, secret_key, token, method=self.METHOD
+                    access_key,
+                    secret_key,
+                    token,
+                    method=self.METHOD,
+                    account_id=account_id,
                 )
         else:
             return None
@@ -747,11 +777,14 @@ class AioAssumeRoleProvider(AssumeRoleProvider):
     async def _resolve_credentials_from_profile(self, profile_name):
         profiles = self._loaded_config.get('profiles', {})
         profile = profiles[profile_name]
-
         if (
             self._has_static_credentials(profile)
             and not self._profile_provider_builder
         ):
+            # This is only here for backwards compatibility. If this provider
+            # isn't given a profile provider builder we still want to be able
+            # to handle the basic static credential case as we would before the
+            # profile provider builder parameter was added.
             return self._resolve_static_credentials_from_profile(profile)
         elif self._has_static_credentials(
             profile
@@ -770,7 +803,6 @@ class AioAssumeRoleProvider(AssumeRoleProvider):
                     error_msg=error_message % profile_name,
                 )
             return credentials
-
         return await self._load_creds_via_assume_role(profile_name)
 
     def _resolve_static_credentials_from_profile(self, profile):
@@ -923,6 +955,7 @@ class AioContainerProvider(ContainerProvider):
             method=self.METHOD,
             expiry_time=_parse_if_needed(creds['expiry_time']),
             refresh_using=fetcher,
+            account_id=creds.get('account_id'),
         )
 
     def _create_fetcher(self, full_uri, *args, **kwargs):
@@ -944,6 +977,7 @@ class AioContainerProvider(ContainerProvider):
                 'secret_key': response['SecretAccessKey'],
                 'token': response['Token'],
                 'expiry_time': response['Expiration'],
+                'account_id': response.get('AccountId'),
             }
 
         return fetch_creds
@@ -971,52 +1005,9 @@ class AioCredentialResolver(CredentialResolver):
         return None
 
 
-class AioSSOCredentialFetcher(AioCachedCredentialFetcher):
-    _UTC_DATE_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
-
-    def __init__(
-        self,
-        start_url,
-        sso_region,
-        role_name,
-        account_id,
-        client_creator,
-        token_loader=None,
-        cache=None,
-        expiry_window_seconds=None,
-        token_provider=None,
-        sso_session_name=None,
-    ):
-        self._client_creator = client_creator
-        self._sso_region = sso_region
-        self._role_name = role_name
-        self._account_id = account_id
-        self._start_url = start_url
-        self._token_loader = token_loader
-        self._token_provider = token_provider
-        self._sso_session_name = sso_session_name
-        super().__init__(cache, expiry_window_seconds)
-
-    def _create_cache_key(self):
-        args = {
-            'roleName': self._role_name,
-            'accountId': self._account_id,
-        }
-        if self._sso_session_name:
-            args['sessionName'] = self._sso_session_name
-        else:
-            args['startUrl'] = self._start_url
-
-        args = json.dumps(args, sort_keys=True, separators=(',', ':'))
-        argument_hash = sha1(args.encode('utf-8')).hexdigest()
-        return self._make_file_safe(argument_hash)
-
-    def _parse_timestamp(self, timestamp_ms):
-        # fromtimestamp expects seconds so: milliseconds / 1000 = seconds
-        timestamp_seconds = timestamp_ms / 1000.0
-        timestamp = datetime.datetime.fromtimestamp(timestamp_seconds, tzutc())
-        return timestamp.strftime(self._UTC_DATE_FORMAT)
-
+class AioSSOCredentialFetcher(
+    SSOCredentialFetcher, AioCachedCredentialFetcher
+):
     async def _get_credentials(self):
         """Get credentials by calling SSO get role credentials."""
         config = Config(
@@ -1050,6 +1041,7 @@ class AioSSOCredentialFetcher(AioCachedCredentialFetcher):
                     'Expiration': self._parse_timestamp(
                         credentials['expiration']
                     ),
+                    'AccountId': self._account_id,
                 },
             }
             return credentials